Stashing work towards interaction editors.
The way that the App view works, other views
are not aware of when their content is hidden or
revealed within toggler interfaces. Unfortunately,
for stuff like CodeMirror, that means that it initializes
without any ClientWidths available, and then when the
toggler button is clicked, you can't grab the event
after the App view does. I think that this should
possibly spawn a top-down event for visibility changing
in app views.

Currently this means that we have mustache and HTML
highlighting, but it isn't active until you
check out the element with Web Developer extensions.
CodeMirror's .refresh() method doesn't work - it's
unclear whether this is really a CodeMirror bug or
some odd browser bug, but it exists in Firefox, Chrome,
and Safari.

This is on the road to addressing #838